feat: Updated docker-compose. Added Dockerfile and docker-compose for NodeJS API. Minor changes in folder names.
