+++ .env
@@ -2,3 +2,4 @@ MONGO_USERNAME=DAS
 MONGO_PASSWORD=sistemas
 MONGO_HOSTNAME=mongo
 MONGO_PORT=27017
+MONGO_DB=OrdinarioDAS
