--- Contenedor D/package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "express": "^4.17.1",
-    "mongodb": "^3.6.3"
-  }
-}
