+++ Contenedor-d/.dockerignore
@@ -1,0 +1,5 @@
+.git
+.gitignore
+README.md
+docker-compose.yml
+node_modules
