--- contenedor-a/mongo-volume/WiredTiger
@@ -1,2 +1,0 @@
-WiredTiger
-WiredTiger 10.0.0: (March 18, 2020)
