--- contenedor-a/mongo-volume/WiredTiger.lock
@@ -1,1 +1,0 @@
-WiredTiger lock file
