--- contenedor-a/mongo-volume/WiredTiger.turtle
@@ -1,6 +1,0 @@
-WiredTiger version string
-WiredTiger 10.0.0: (March 18, 2020)
-WiredTiger version
-major=10,minor=0,patch=0
-file:WiredTiger.wt
-access_pattern_hint=none,allocation_size=4KB,app_metadata=,assert=(commit_timestamp=none,durable_timestamp=none,read_timestamp=none),block_allocation=best,block_compressor=,cache_resident=false,checksum=uncompressed,collator=,columns=,dictionary=0,encryption=(keyid=,name=),format=btree,huffman_key=,huffman_value=,id=0,ignore_in_memory_cache_size=false,internal_item_max=0,internal_key_max=0,internal_key_truncate=true,internal_page_max=4KB,key_format=S,key_gap=10,leaf_item_max=0,leaf_key_max=0,leaf_page_max=32KB,leaf_value_max=0,log=(enabled=true),memory_page_image_max=0,memory_page_max=5MB,os_cache_dirty_max=0,os_cache_max=0,prefix_compression=false,prefix_compression_min=4,split_deepen_min_child=0,split_deepen_per_child=0,split_pct=90,value_format=S,version=(major=1,minor=1),checkpoint=(WiredTigerCheckpoint.6=(addr="018081e3675edb8181e45c13e2a18281e42b3964f0808080e2efc0e22fc0",order=6,time=1606753871,size=24576,newest_start_durable_ts=0,oldest_start_ts=0,newest_txn=5,newest_stop_durable_ts=0,newest_stop_ts=-1,newest_stop_txn=-11,prepare=0,write_gen=14)),checkpoint_backup_info=,checkpoint_lsn=(2,6400)
