+++ docker-compose.yml
@@ -34,6 +34,27 @@ services:
       - 'mongo'
     depends_on:
       - 'mongo'
+  # Contenedor D.
+  nodeapi:
+    container_name: nodeapi
+    image: nodejs:latest
+    restart: unless-stopped
+    build:
+      context: ./Contenedor D/.
+      dockerfile: Dockerfile
+    env_file: .env
+    environment:
+      MONGO_USERNAME: ${MONGO_USERNAME}
+      MONGO_PASSWORD: ${MONGO_PASSWORD}
+      MONGO_HOSTNAME: ${MONGO_HOSTNAME}
+      MONGO_PORT: ${MONGO_PORT}
+      MONGO_DB: ${MONGO_DB} 
+    ports:
+      - "1337:1337"
+    links:
+      - 'mongo'
+    depends_on:
+      - 'mongo'
   # Contenedor F.
   services:
     container_name: apache
